[wpigui] Add GetDistSquared and MaxFit functions
These utility functions were moved from halsim_gui.
